Keep ruff 0.16 green: allow formula typography, adopt the two new lint idioms
ruff 0.16.2 (venv upgrade) added RUF001/002/003 confusable checks that
flag the deliberate ×/− in calc formula docstrings — allowed explicitly
in config rather than degrading the notation — plus RUF005/RUF046,
adopted as written (tuple unpacking, redundant int()). No behavior
change; 218 tests stay green.

+++ Ms-Agent/app/services/matching_service.py
@@ -404,7 +404,7 @@ def calculate_match_score(
     if evaluable:
         total_weight = sum(c.weight for c in evaluable)
         raw = sum(c.score * c.weight for c in evaluable) / total_weight
-        match_score = int(round(raw))
+        match_score = round(raw)
     else:
         match_score = 0
